refactor: separate modules for functions draw figures and screenshot
